# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nexo API",
        "description": "Fiscal centralizing API for integration between sales systems, government and accounting platforms.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        },
        {
            "url": "https://nexo.test",
            "description": "Herd Local Server (HTTPS)"
        },
        {
            "url": "https://backendnexo.shop",
            "description": "Mounted Server"
        }
    ],
    "paths": {
...
        "/api/buyers": {
            "get": {
                "tags": [
                    "Buyers"
                ],
                "summary": "List buyers of an organization",
                "description": "Returns a paginated list of buyers belonging to the given organization. Admin endpoint, no authentication required.",
                "operationId": "indexBuyers",
                "parameters": [
                    {
                        "name": "organization_id",
                        "in": "query",
                        "description": "ID of the organization to filter buyers by.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of results per page. Defaults to 15.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15,
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Buyers retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Buyers retrieved successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "organization_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Empresa ABC"
                                                    },
                                                    "document_number": {
                                                        "type": "string",
                                                        "example": "12345678"
                                                    },
                                                    "active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error — organization_id missing or invalid."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Buyers"
                ],
                "summary": "Create a new buyer",
                "description": "Creates a buyer in the catalog. The document_number must be unique within the organization. Admin endpoint, no authentication required.",
                "operationId": "storeBuyer",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "organization_id",
                                    "name"
                                ],
                                "properties": {
                                    "organization_id": {
                                        "description": "ID of the organization.",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "description": "Buyer display name.",
                                        "type": "string",
                                        "example": "Empresa ABC"
                                    },
                                    "document_number": {
                                        "description": "Optional document number (8-20 digits, country-agnostic).",
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Buyer created successfully."
                    },
                    "409": {
                        "description": "Conflict — buyer with same document number already exists in this organization."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/buyers/{id}": {
            "get": {
                "tags": [
                    "Buyers"
                ],
                "summary": "Get a buyer by ID",
                "description": "Returns a single buyer. Admin endpoint, no authentication required.",
                "operationId": "showBuyer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the buyer.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Buyer retrieved successfully."
                    },
                    "404": {
                        "description": "Buyer not found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Buyers"
                ],
                "summary": "Update a buyer",
                "description": "Updates name and/or document_number. The document_number must remain unique within the organization. Admin endpoint, no authentication required.",
                "operationId": "updateBuyer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the buyer to update.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "New name.",
                                        "type": "string",
                                        "example": "Empresa Renombrada"
                                    },
                                    "document_number": {
                                        "description": "New document number, or null to clear.",
                                        "type": "string",
                                        "example": "87654321"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Buyer updated successfully."
                    },
                    "404": {
                        "description": "Buyer not found."
                    },
                    "409": {
                        "description": "Conflict — buyer with same document number already exists in this organization."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Buyers"
                ],
                "summary": "Delete a buyer",
                "description": "Permanently removes a buyer from the catalog. Admin endpoint, no authentication required.",
                "operationId": "destroyBuyer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the buyer to delete.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Buyer deleted successfully."
                    },
                    "404": {
                        "description": "Buyer not found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "agentToken": {
                "type": "http",
                "description": "Bearer token issued by POST /api/agent/token. Used by the nexo-agent daemon to authenticate all agent endpoints.",
                "scheme": "bearer"
            },
            "bearerAuth": {
                "type": "http",
                "description": "Bearer JWT issued by POST /api/auth/login. Pass as: Authorization: Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "connectorToken": {
                "type": "http",
                "description": "Bearer token issued when a Connector is created. Only visible at creation time.",
                "scheme": "bearer"
            }
        }
    },
...
}